+++ jano_web/src/main/java/ec/com/uce/jano/web/beans/RecaudacionIngresoBean.java
@@ -28,8 +28,8 @@ public class RecaudacionIngresoBean implements Serializable {
 	private String observacion;
 	private String beneficiario;
 	private double valor;
-	private Integer facultad;
-	private Integer dependencia;
+	private Long facultad;
+	private Long dependencia;
 	private Long idAfectacion;
 	private String cur;
 
@@ -126,7 +126,7 @@ public class RecaudacionIngresoBean implements Serializable {
 	/**
 	 * @return the facultad
 	 */
-	public Integer getFacultad() {
+	public Long getFacultad() {
 		return facultad;
 	}
 
@@ -134,14 +134,14 @@ public class RecaudacionIngresoBean implements Serializable {
 	 * @param facultad
 	 *            the facultad to set
 	 */
-	public void setFacultad(Integer facultad) {
+	public void setFacultad(Long facultad) {
 		this.facultad = facultad;
 	}
 
 	/**
 	 * @return the dependencia
 	 */
-	public Integer getDependencia() {
+	public Long getDependencia() {
 		return dependencia;
 	}
 
@@ -149,7 +149,7 @@ public class RecaudacionIngresoBean implements Serializable {
 	 * @param dependencia
 	 *            the dependencia to set
 	 */
-	public void setDependencia(Integer dependencia) {
+	public void setDependencia(Long dependencia) {
 		this.dependencia = dependencia;
 	}
